tests: add rudimentary tests for resolveIPs
This mocks net.LookupIP with a wrapper type. In future, this might
be easier to accomplish with the help of the standard library (see
discussion in golang/go#12503).

With muety/caddy-remote-host#1 in the pipeline, I might consider
swapping the stdlib resolver with a full fledged DNS client
(github.com/miekg/dns), as the stdlib resolver does not return
TTL values.